requirements.txt
# 核心依赖
mermaid-py>=0.2.0
matplotlib>=3.7.0
pandas>=1.5.0
seaborn>=0.12.0
numpy>=1.23.0
pillow>=9.5.0
pydantic>=2.0.0

# 服务相关依赖
fastmcp>=0.1.0
six>=1.16.0
uvicorn>=0.22.0

# 开发工具
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# 系统兼容性工具
platformdirs>=3.2.0